rename entity

+++ Base.WS.API/JS.Base.WS.API/Base/Audit.cs
@@ -9,10 +9,10 @@ namespace JS.Base.WS.API.Base
     public abstract class Audit : IAudit
     {
         public DateTime? CreationTime { get; set; }
-        public int? CreatorUserId { get; set; }
+        public long? CreatorUserId { get; set; }
         public DateTime? LastModificationTime { get; set; }
-        public int? LastModifierUserId { get; set; }
-        public int? DeleterUserId { get; set; }
+        public long? LastModifierUserId { get; set; }
+        public long? DeleterUserId { get; set; }
         public DateTime? DeletionTime { get; set; }
         public bool IsActive { get; set; }
         public bool IsDeleted { get; set; }
