+++ Base.WS.API/JS.Base.WS.API/Base/IAudit.cs
@@ -9,10 +9,10 @@ namespace JS.Base.WS.API.Base
     interface IAudit
     {
         DateTime? CreationTime { get; set; }
-        int? CreatorUserId { get; set; }
+        long? CreatorUserId { get; set; }
         DateTime? LastModificationTime { get; set; }
-        int? LastModifierUserId { get; set; }
-        int? DeleterUserId { get; set; }
+        long? LastModifierUserId { get; set; }
+        long? DeleterUserId { get; set; }
         DateTime? DeletionTime { get; set; }
         bool IsActive { get; set; }
         bool IsDeleted { get; set; }
