+++ Base.WS.API/JS.Base.WS.API/Models/Authorization/User.cs
@@ -12,14 +12,14 @@ namespace JS.Base.WS.API.Models.Authorization
     public class User: Audit
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string UserName { get; set; }
         public string Password { get; set; }
         public string Name { get; set; }
         public string Surname { get; set; }
         public string EmailAddress { get; set; }
         public int StatusId { get; set; }
-        public int? PersonId { get; set; }
+        public long? PersonId { get; set; }
 
         [ForeignKey("StatusId")]
         public virtual UserStatus UserStatus { get; set; }
